finetune/requirements.txt
# MindWall Fine-Tuning Dependencies
# Developed by Pradyumn Tandon (https://pradyumntandon.com) at VRIP7 (https://vrip7.com)
# Run locally on 8GB+ VRAM GPU

# Core
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
trl>=0.7.0
transformers>=4.38.0
datasets>=2.17.0
accelerate>=0.27.0
peft>=0.9.0
bitsandbytes>=0.42.0
sentencepiece>=0.1.99

# Evaluation
scikit-learn>=1.4.0
rouge-score>=0.1.2
nltk>=3.8.1

# Export
llama-cpp-python>=0.2.50

# Data preparation
pandas>=2.2.0
requests>=2.31.0
tqdm>=4.66.0
pyyaml>=6.0.1

# Utilities
torch>=2.1.0
numpy>=1.26.0
